Code refactor
Added controllers and services and added moved the code inside routes to these. Routes now just map a particular HTTP request with the appropriate controller that handles them.
